--- a/packages/@angular/cli/models/webpack-configs/production.ts
+++ b/packages/@angular/cli/models/webpack-configs/production.ts
@@ -152,6 +152,7 @@ export function getProdConfig(wco: WebpackConfigOptions) {
       new UglifyJSPlugin({
         sourceMap: buildOptions.sourcemaps,
         parallel: true,
+        cache: true,
         uglifyOptions: {
           ecma: wco.supportES2015 ? 6 : 5,
           warnings: buildOptions.verbose,