This PR adds cache support for UglifyJS and Build Optimizer, and executes Build Optimizer in parallel.

Benchmarks for `ng build --prod`:
- new project
  - before: 16270 ms
  - after (no cache): 17283 ms (+6%)
  - after (with cache): 7040 ms (-56%)
- big project (https://github.com/filipesilva/angular-cli-test-repo)
  - before: 629958 ms
  - after (no cache): 571516 ms (-9%)
  - after (with cache): 456342 ms (-28%)