@Toxicable 2018-01-19
@filipesilva 
Looking at https://github.com/amireh/happypack#threads-number
Could you try bench-marking it with threads set to 4?