@Toxicable 2018-01-19
Also since we're adding a cache could we see some RAM benchmarks?