@clydin 2018-01-19
Using the number of cpu cores would probably be a good idea.  Too many threads will end up hurting performance due to excessive context switching and too little will limit the potential performance improvement.