@Toxicable 2018-01-19
Since according to the readme there's diminishing returns after 4, I think something like this would be best 
```
const cpus = require('os').cpus().length;
const happyCores = cpus > 4 ? 4 : cpus;
```