@clydin 2018-01-20
Most likely because the tests were run on a quad core system which would make 4 the best number to use.