@Toxicable 2018-01-20
@clydin Possibly, I think it would be worth doing our own analysis either way since this could be a big factor on the final speed