@filipesilva 2018-01-22
I tried adding the CPU config to the same that `uglifyjs-webpack-plugin` uses, which in my case will be 7:
```
    extraPlugins.push(new HappyPack({
      verbose: false,
      threads: require('os').cpus().length - 1,
```

Then I ran cleared node_modules and ran 2 prod builds in a row:
- default number of threads
  - no cache: 548029 ms (~9.1 minutes)
  - with cache: 479455 ms (~7.9 minutes)
- 7 threads
  - no cache: 559157 ms (~9.3 minutes)
  - with cache: 454367 ms (~7.5 minutes)

I also tried a new project:
- default number of threads
  - no cache: 15617 ms (~16 seconds)
  - with cache: 6350 ms (~6 seconds)
- 7 threads
  - no cache: 16393 ms (~16 seconds)
  - with cache: 7527 ms (~8 seconds)

Doesn't seem a big difference overall. But I think making it variable it might be worth it for some environments with less than 3 CPUs to avoid spawning more threads than needed, but probably defaulting to 1 if there's just one CPU.

PR is amended accordingly.