@filipesilva 2018-01-23
I removed the parallelization to reduce risk since it looks like cache is what provided the bigger benefit.